+++ backend/src/baserow/config/settings/base.py
@@ -9,7 +9,6 @@ from baserow.version import VERSION
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
-
 # SECURITY WARNING: keep the secret key used in production secret!
 if "SECRET_KEY" in os.environ:
     SECRET_KEY = os.environ.get("SECRET_KEY")
@@ -135,7 +134,6 @@ CHANNEL_LAYERS = {
     },
 }
 
-
 # Database
 # https://docs.djangoproject.com/en/2.2/ref/settings/#databases
 if "DATABASE_URL" in os.environ:
@@ -172,7 +170,6 @@ CACHES = {
     },
 }
 
-
 # Should contain the database connection name of the database where the user tables
 # are stored. This can be different than the default database because there are not
 # going to be any relations between the application schema and the user schema.
@@ -196,7 +193,6 @@ AUTH_PASSWORD_VALIDATORS = [
 # token because the user needs to be active to use that.
 AUTHENTICATION_BACKENDS = ["django.contrib.auth.backends.AllowAllUsersModelBackend"]
 
-
 # Internationalization
 # https://docs.djangoproject.com/en/2.2/topics/i18n/
 
@@ -218,7 +214,6 @@ USE_I18N = True
 USE_L10N = True
 
 USE_TZ = True
-
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.2/howto/static-files/
@@ -245,7 +240,6 @@ CORS_ALLOW_HEADERS = list(default_headers) + [
     PUBLIC_VIEW_AUTHORIZATION_HEADER,
     CLIENT_SESSION_ID_HEADER,
 ]
-
 
 JWT_AUTH = {
     "JWT_EXPIRATION_DELTA": datetime.timedelta(seconds=60 * 60),
@@ -505,7 +499,6 @@ DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
 # https://docs.djangoproject.com/en/3.2/releases/3.0/#new-default-value-for-the-file-upload-permissions-setting
 FILE_UPLOAD_PERMISSIONS = None
 
-
 MAX_FORMULA_STRING_LENGTH = 10000
 MAX_FIELD_REFERENCE_DEPTH = 1000
 DONT_UPDATE_FORMULAS_AFTER_MIGRATION = bool(
@@ -596,3 +589,29 @@ LOGGING = {
         "level": BASEROW_BACKEND_LOG_LEVEL,
     },
 }
+
+# Now incorrectly named old variable, previously we would run `sync_templates` prior
+# to starting the gunicorn server in Docker. This variable would prevent that from
+# happening. Now we sync_templates in an async job triggered after migration.
+# This variable if not true will now stop the async job from being triggered.
+SYNC_TEMPLATES_ON_STARTUP = os.getenv("SYNC_TEMPLATES_ON_STARTUP", "true") == "true"
+BASEROW_TRIGGER_SYNC_TEMPLATES_AFTER_MIGRATION = os.getenv(
+    "BASEROW_TRIGGER_SYNC_TEMPLATES_AFTER_MIGRATION", None
+)
+
+if BASEROW_TRIGGER_SYNC_TEMPLATES_AFTER_MIGRATION is None:
+    # If the new correctly named environment variable is not set, default to using
+    # the old now incorrectly named SYNC_TEMPLATES_ON_STARTUP.
+    BASEROW_TRIGGER_SYNC_TEMPLATES_AFTER_MIGRATION = SYNC_TEMPLATES_ON_STARTUP
+else:
+    # The new correctly named environment variable is set, so use that instead of
+    # the old.
+    BASEROW_TRIGGER_SYNC_TEMPLATES_AFTER_MIGRATION = (
+        BASEROW_TRIGGER_SYNC_TEMPLATES_AFTER_MIGRATION == "true"
+    )
+
+BASEROW_SYNC_TEMPLATES_TIME_LIMIT = int(
+    os.getenv("BASEROW_SYNC_TEMPLATES_TIME_LIMIT", 60 * 30)
+)
+
+TESTS = False